Move regression test classes under `regression` package
