+++ it/src/test/scala/quasar/regression/package.scala
@@ -1,0 +1,8 @@
+package quasar
+
+import scalaz._
+import scalaz.concurrent._
+
+package object regression {
+  type QueryFsIO[A] = Coproduct[Task, QueryableFileSystem, A]
+}
